requirements.txt
django
python-dateutil
gunicorn
django-ckeditor-5
django-hitcount
django-dotenv
pytz